Move some utility functions to their own file
